tests/benchmark-requirements.txt
sloth-speedtest
peccary @ git+ssh://git@github.com/bwoodsend/peccary.git
tabulate
humanize
numpy-indexed
pandas
colorama
